Initial version of app
This brings together all of the elements and gives us an app the allows us to add new expense items which are automatically saved in UserDefaults and displayed in a list in the main view. The items can be deleted from the list. Individual row items employ some basic layout modification to display the name and time in a VStack with the amount appearing to the right due to the VStack and Text views being wrapped in an HStack.